Move unit tests to MULANGT10BaseCodopTest
